checkPermission is added && update file's

+++ backend/Routes/Roadmap/addProgressState.js
@@ -1,11 +1,21 @@
 const router = require("express").Router();
 const pool = require("../../Database/db");
 const authorization = require("../../middleware/authorization");
-
+const checkPermission = require("../../middleware/checkPermissions");
 router.post("/student/state", authorization, async (req, res) => {
   try {
     const { topic_id, topic_level, state_id } = req.body;
     const studentId = req.user.userId;
+    const roleId = req.user.roleId;
+    //permission
+    const hasAccess = await checkPermission(
+      studentId,
+      "addProgressState",
+      roleId
+    );
+    if (!hasAccess) {
+      return res.status(403).json("Access denied");
+    }
     // Verify before adding(true or false)
     const verify = await pool.query(
       ` SELECT EXISTS (
@@ -13,25 +23,23 @@ router.post("/student/state", authorization, async (req, res) => {
       [topic_id, topic_level, studentId]
     );
     // * Update state
-    if(verify.rows[0].exists){
+    if (verify.rows[0].exists) {
       try {
         const updateResponse = await pool.query(
           `UPDATE Progress_Status
              SET topic_id=$1, topic_level=$2, state_id=$4
              WHERE topic_id=$1 AND topic_level=$2 AND student_id=$3`,
           [topic_id, topic_level, studentId, state_id]
-          
         );
-        return res.status(200).json({ status: "success: topic was updated state" });
-        
+        return res
+          .status(200)
+          .json({ status: "success: topic was updated state" });
       } catch (error) {
-            console.error(err);
-            res
-              .status(500)
-              .json({ status: "error", message: "An error occurred" });
+        console.error(err);
+        res.status(500).json({ status: "error", message: "An error occurred" });
       }
     }
-    
+
     const newRodmapInfo = await pool.query(
       "INSERT INTO Progress_Status (topic_id, topic_level, student_id,state_id) VALUES ($1, $2, $3,$4) ",
       [topic_id, topic_level, studentId, state_id]
@@ -44,23 +52,37 @@ router.post("/student/state", authorization, async (req, res) => {
   }
 });
 
-// reset 
-router.delete("/student/reset/:topicId/:level", authorization, async (req, res) => {
-  try {
-    const topic_id = req.params.topicId;
-    const level=req.params.level;
-    const studentId = req.user.userId;
-    const newRodmapInfo = await pool.query(
-      "DELETE FROM Progress_Status WHERE topic_id = $1 AND student_id = $2 AND topic_level=$3",
-      [topic_id, studentId,level]
-    );
+// reset
+router.delete(
+  "/student/reset/:topicId/:level",
+  authorization,
+  async (req, res) => {
+    try {
+      const topic_id = req.params.topicId;
+      const level = req.params.level;
+      const studentId = req.user.userId;
+      const roleId = req.user.roleId;
+      //permission
+      const hasAccess = await checkPermission(
+        studentId,
+        "addProgressState",
+        roleId
+      );
+      if (!hasAccess) {
+        return res.status(403).json("Access denied");
+      }
+      const newRodmapInfo = await pool.query(
+        "DELETE FROM Progress_Status WHERE topic_id = $1 AND student_id = $2 AND topic_level=$3",
+        [topic_id, studentId, level]
+      );
 
-    res.status(204).json({ status: "success: topic was reset" });
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ status: "error", message: "An error occurred" });
+      res.status(204).json({ status: "success: topic was reset" });
+    } catch (err) {
+      console.error(err);
+      res.status(500).json({ status: "error", message: "An error occurred" });
+    }
   }
-});
+);
 
 module.exports = router;
